pclientd: split into separate crate
