+++ pclientd/Cargo.toml
@@ -1,0 +1,50 @@
+[package]
+name = "pclientd"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+# Workspace dependencies
+penumbra-proto = { path = "../proto", features = ["rpc"] }
+penumbra-chain = { path = "../chain" }
+penumbra-crypto = { path = "../crypto" }
+penumbra-tct = { path = "../tct" }
+penumbra-transaction = { path = "../transaction" }
+penumbra-custody = {path = "../custody"}
+penumbra-view = { path = "../view"}
+# TODO: replace by a penumbra-app
+penumbra-component = { path = "../component" }
+
+sqlx = { version = "0.6", features = [ "runtime-tokio-rustls", "offline", "sqlite" ] }
+tokio = { version = "1.22", features = ["full"]}
+tokio-stream = { version =  "0.1.8", features = ["sync"] }
+anyhow = "1"
+rand_core = { version = "0.6.3", features = ["getrandom"] }
+rand = "0.8"
+serde_json = "1"
+serde = { version = "1", features = ["derive"] }
+serde_with = { version = "1.11", features = ["hex"] }
+tracing = "0.1"
+tracing-subscriber = "0.2"
+tonic = "0.8.1"
+tonic-web = "0.4.0"
+bytes = { version = "1", features = ["serde"] }
+prost = "0.11"
+futures = "0.3"
+hex = "0.4"
+metrics = "0.19.0"
+async-stream = "0.2"
+parking_lot = "0.12"
+clap = { version = "3", features = ["derive"] }
+camino = "1"
+async-trait = "0.1"
+tendermint = "0.29.0"
+sha2 = "0.10.1"
+toml = "0.5"
+ed25519-consensus = "2.1"
+
+[build-dependencies]
+vergen = "5"
+
