+++ pclientd/build.rs
@@ -1,0 +1,5 @@
+use vergen::{vergen, Config};
+
+fn main() {
+    vergen(Config::default()).unwrap();
+}
